Remove underscores and periods in XML identifiers
BUG: The code is supposed to generate unexported types by default.
All identifiers that are not Go builtin types are converted to PascalCase.
This means generated types are now always exported.

The code assumed all XML identifiers were lowercase dash-delimited strings,
so it was not reliably producing unexported types anyway.

If a type was not defined in the XSD, assign a default value of string.